Rename C_scalar_field to B_scalar_field module
All files and imports under openwave/xperiments/C_scalar_field were renamed to openwave/xperiments/B_scalar_field. References in code were updated accordingly to reflect the new module name.
