+++ openwave/xperiments/B_scalar_field/__init__.py
@@ -1,5 +1,5 @@
 """
-C/ SCALAR-FIELD METHOD (analytical tools)
+B/ SCALAR-FIELD METHOD (analytical tools)
 
 Experiments exploring wave dynamics using scalar-field method simulation.
 """
