+++ openwave/xperiments/B_scalar_field/plots/charge_levels.png
