+++ openwave/xperiments/B_scalar_field/plots/probe_values.png
