+++ openwave/xperiments/B_scalar_field/research/docs/01b_WAVE_FIELD_properties.md
