+++ openwave/xperiments/B_scalar_field/research/docs/06_PHOTON_HEAT.md
