+++ openwave/xperiments/B_scalar_field/research/docs/13_wave_center.md
