+++ openwave/xperiments/B_scalar_field/research/docs/stable_waves.md
