+++ openwave/xperiments/B_scalar_field/research/docs/support_material/images/laplacian18-1M.png
