+++ openwave/xperiments/B_scalar_field/research/docs/support_material/images/laplacian6-1M.png
