+++ openwave/xperiments/B_scalar_field/research/docs/support_material/images/spherical_symmetry.jpeg
