+++ openwave/xperiments/B_scalar_field/research/docs/support_material/images/standing_waves.png
