+++ openwave/xperiments/B_scalar_field/research/docs/support_material/taichi_autodiff.md
