+++ openwave/xperiments/B_scalar_field/research/docs/support_material/wave_exponential.md
