+++ openwave/xperiments/B_scalar_field/research/docs/support_material/wave_vs_heat_equation.md
